+++ DataFormats/BTauReco/interface/TauImpactParameterInfo.h
@@ -8,7 +8,6 @@
 #include "DataFormats/BTauReco/interface/TauImpactParameterInfoFwd.h"
 #include "DataFormats/GeometryCommonDetAlgo/interface/Measurement1D.h"
 #include "DataFormats/BTauReco/interface/IsolatedTauTagInfo.h"
-#include "DataFormats/BTauReco/interface/IsolatedTauTagInfoFwd.h"
 #include "DataFormats/BTauReco/interface/TrackTauImpactParameterAssociation.h"
 
 using namespace std;
@@ -29,8 +28,8 @@ namespace reco {
     
     virtual TauImpactParameterInfo* clone() const { return new TauImpactParameterInfo( * this ); }
     
-    double discriminator(double,double,double,bool,bool) const;
-    double discriminator() const;
+    float discriminator(double,double,double,bool,bool) const;
+    float discriminator() const;
     
     const TauImpactParameterTrackData* getTrackData(reco::TrackRef) const;
     void 	         storeTrackData(reco::TrackRef,const TauImpactParameterTrackData&);
